+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Docker image dependencies that are DNAnexus platform files are included in applet's bundledDepends
 * Adds `-instanceTypeSelection` compiler option to allow disabling compile-type instance type selection
 * Adds `-defaultInstanceType` option
+* Adds support for new London region, `aws:eu-west-2-g`
 
 ## 2.4.10 2021-08-16